refactor: collapse daemon action facade types
The daemon action proxy and its public mapped type now live in the same module. LoadedWorkspace no longer names actions as a special field, and the daemon action type drops arrays to better match runtime discovery.
